compress/pyproject.toml
[project]
name = "compress"
version = "0.1.0"
description = "Compress images to a target size (~1MB by default)"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.2.0",
]
